fix(execution): use int64_t row counters and always show them in --explain-analyze
rows_in/rows_out were int, risking overflow on large tables, and were
suppressed whenever a stage genuinely processed zero rows — hiding a
real signal in EXPLAIN ANALYZE output.

+++ src/planner/plan_node.h
@@ -8,8 +8,8 @@
 using Row = std::vector<Value>;
 
 struct NodeStats {
-    int rows_in  = 0;
-    int rows_out = 0;
+    int64_t rows_in  = 0;
+    int64_t rows_out = 0;
     double elapsed_us = 0.0;
 };
